+++ apps/web/app/(use-page-wrapper)/apps/page.tsx
@@ -1,0 +1,27 @@
+import { withAppDirSsr } from "app/WithAppDirSsr";
+import { PageProps } from "app/_types";
+import { _generateMetadata } from "app/_utils";
+import { cookies, headers } from "next/headers";
+
+import { getServerSideProps } from "@lib/apps/getServerSideProps";
+import { buildLegacyCtx } from "@lib/buildLegacyCtx";
+
+import AppsPage from "~/apps/apps-view";
+
+export const generateMetadata = async () => {
+  return await _generateMetadata(
+    (t) => t("app_store"),
+    (t) => t("app_store_description")
+  );
+};
+
+const getData = withAppDirSsr(getServerSideProps);
+
+const ServerPage = async ({ params, searchParams }: PageProps) => {
+  const context = buildLegacyCtx(headers(), cookies(), params, searchParams);
+
+  const props = await getData(context);
+  return <AppsPage {...props} />;
+};
+
+export default ServerPage;
