+++ apps/web/app/(use-page-wrapper)/availability/troubleshoot/page.tsx
@@ -1,0 +1,12 @@
+import { _generateMetadata } from "app/_utils";
+
+import Troubleshoot from "~/availability/troubleshoot/troubleshoot-view";
+
+export const generateMetadata = async () => {
+  return await _generateMetadata(
+    (t) => t("troubleshoot"),
+    (t) => t("troubleshoot_availability")
+  );
+};
+
+export default Troubleshoot;
