+++ apps/web/app/(use-page-wrapper)/insights/page.tsx
