+++ apps/web/app/(use-page-wrapper)/insights/virtual-queues/page.tsx
