--- apps/web/app/insights/layout.tsx
@@ -1,5 +1,0 @@
-import { WithLayout } from "app/layoutHOC";
-
-import { getInsightsLayout } from "~/insights/layout";
-
-export default WithLayout({ getLayout: getInsightsLayout });
