--- apps/web/app/signup/page.tsx
@@ -1,22 +1,0 @@
-import { withAppDirSsr } from "app/WithAppDirSsr";
-import { _generateMetadata } from "app/_utils";
-import { WithLayout } from "app/layoutHOC";
-
-import { getServerSideProps } from "@lib/signup/getServerSideProps";
-
-import type { SignupProps } from "~/signup-view";
-import Signup from "~/signup-view";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("sign_up"),
-    (t) => t("sign_up")
-  );
-
-const getData = withAppDirSsr<SignupProps>(getServerSideProps);
-
-export default WithLayout({
-  Page: Signup,
-  getLayout: null,
-  getData,
-})<"P">;
