+++ apps/web/modules/availability/troubleshoot/troubleshoot-view.tsx
@@ -26,6 +26,5 @@ function TroubleshooterPage() {
     </>
   );
 }
-export { getLayout } from "@calcom/features/troubleshooter/layout";
 
 export default TroubleshooterPage;
